+++ OpenGLEngine.xcodeproj/xcshareddata/xcschemes/OpenGLEngine.xcscheme
@@ -36,7 +36,7 @@
       selectedLauncherIdentifier = "Xcode.DebuggerFoundation.Launcher.LLDB"
       launchStyle = "0"
       useCustomWorkingDirectory = "YES"
-      customWorkingDirectory = "$PROJECT_DIR"
+      customWorkingDirectory = "/Users/s.k/Documents/Develop/OpenGL/OpenGLEngine/OpenGLEngine"
       ignoresPersistentStateOnLaunch = "NO"
       debugDocumentVersioning = "YES"
       debugServiceExtension = "internal"
