+++ OpenGLEngine/Header/environment.h
@@ -1,0 +1,3 @@
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wdocumentation"
+#pragma GCC diagnostic ignored "-Wdeprecated-declarations"
