+++ OpenGLEngine/Shader/point.vert
@@ -2,5 +2,5 @@
 in vec4 position;
 
 void main(){
-    gl_position = position;
+    gl_Position = position;
 }
